requirements-knowledge-base.txt
llama-index-core>=0.12.0
llama-index-llms-openai>=0.3.0
llama-index-llms-openai-like>=0.3.0
llama-index-embeddings-openai>=0.3.0
llama-index-embeddings-huggingface>=0.5.0
sentence-transformers>=3.0.0
